requirements.txt
pandas~=1.2.4
numpy~=1.20.3
requests~=2.25.1
oss2~=2.15.0
python-jenkins~=1.7.0
jmespath~=0.10.0
py~=1.10.0
pbr~=5.8.0
setuptools~=57.0.0
six~=1.16.0
virtualenv~=20.4.7
pip~=21.1.2
cryptography~=35.0.0
lxml~=4.6.3
tox~=3.23.1
filelock~=3.0.12
pluggy~=0.13.1
toml~=0.10.2
packaging~=20.9
pytest~=6.2.5
pytz~=2021.1
MarkupSafe~=2.0.1
Werkzeug~=2.0.1
Jinja2~=3.0.1
click~=8.0.1
itsdangerous~=2.0.1
cffi~=1.15.0
openpyxl~=3.0.7
python-dateutil~=2.8.1
certifi~=2021.5.30
chardet~=4.0.0
codecov~=2.1.11
urllib3~=1.26.5
coverage~=5.5
deepdiff~=5.5.0
Pygments~=2.9.0
idna~=2.10
termcolor~=1.1.0
rootpath~=0.1.1
greenlet~=1.1.2
gevent~=21.8.0
blessings~=1.7
pyparsing~=2.4.7
distlib~=0.3.2
appdirs~=1.4.4
humanfriendly~=9.2
coloredlogs~=15.0.1
Flask~=2.0.1
zipp~=3.4.1
cos-python-sdk-v5~=1.9.11
mysql-connector~=2.2.9
flask-caching~=1.10.1
gunicorn~=20.1.0
requests_toolbelt~=0.9.1
jsonpath~=0.82